+++ package.json
@@ -120,6 +120,7 @@
     "emotion": "10.0.27",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
+    "esbuild": "^0.8.30",
     "eslint": "^7.11.0",
     "eslint-config-airbnb-typescript": "^12.0.0",
     "eslint-config-prettier": "^6.0.0",
