--- packages/paste-core/components/alert/tsconfig.build.json
@@ -1,45 +1,0 @@
-{
-  "extends": "../../../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "./src"
-  },
-  "include": [
-    "src/**/*"
-  ],
-  "references": [
-    {
-      "path": "../../primitives/box"
-    },
-    {
-      "path": "../button"
-    },
-    {
-      "path": "../../../paste-design-tokens"
-    },
-    {
-      "path": "../../../paste-icons"
-    },
-    {
-      "path": "../../layout/media-object"
-    },
-    {
-      "path": "../spinner"
-    },
-    {
-      "path": "../../../paste-style-props"
-    },
-    {
-      "path": "../../../paste-libraries/styling"
-    },
-    {
-      "path": "../../../paste-theme"
-    },
-    {
-      "path": "../../../paste-types"
-    },
-    {
-      "path": "../../../paste-libraries/uid"
-    }
-  ]
-}
