+++ packages/paste-core/components/button/build.js
@@ -1,0 +1,3 @@
+const {build} = require('../../../../tools/build/esbuild');
+
+build(require('./package.json'));
