+++ packages/paste-core/components/button/src/types.ts
@@ -1,5 +1,3 @@
-import * as PropTypes from 'prop-types';
-
 type ButtonTypes = 'submit' | 'button' | 'reset';
 export type ButtonSizes = 'small' | 'default' | 'icon' | 'icon_small' | 'reset';
 export type ButtonVariants =
@@ -21,28 +19,6 @@ export interface ButtonContentsProps {
   variant?: ButtonVariants;
 }
 
-export const DirectButtonPropTypes = {
-  as: PropTypes.string as any,
-  fullWidth: PropTypes.bool,
-  href: PropTypes.string,
-  size: PropTypes.oneOf(['small', 'default', 'icon', 'icon_small', 'reset']).isRequired as any,
-  tabIndex: PropTypes.oneOf([0, -1]) as any,
-  type: PropTypes.oneOf(['submit', 'button', 'reset']) as any,
-  disabled: PropTypes.bool,
-  buttonState: PropTypes.oneOf(['disabled', 'loading', 'default']).isRequired as any,
-  variant: PropTypes.oneOf([
-    'primary',
-    'secondary',
-    'destructive',
-    'destructive_link',
-    'destructive_secondary',
-    'link',
-    'inverse_link',
-    'inverse',
-    'reset',
-  ]) as any,
-};
-
 export interface DirectButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   size: ButtonSizes;
   children: React.ReactNode;
@@ -55,28 +31,6 @@ export interface DirectButtonProps extends React.ButtonHTMLAttributes<HTMLButton
   type?: ButtonTypes;
   buttonState: ButtonStates;
 }
-
-export const ButtonPropTypes = {
-  as: PropTypes.string as any,
-  fullWidth: PropTypes.bool,
-  href: PropTypes.string,
-  size: PropTypes.oneOf(['small', 'default', 'icon', 'icon_small', 'reset']) as any,
-  tabIndex: PropTypes.oneOf([0, -1]) as any,
-  type: PropTypes.oneOf(['submit', 'button', 'reset']) as any,
-  disabled: PropTypes.bool,
-  loading: PropTypes.bool,
-  variant: PropTypes.oneOf([
-    'primary',
-    'secondary',
-    'destructive',
-    'destructive_link',
-    'destructive_secondary',
-    'link',
-    'inverse_link',
-    'inverse',
-    'reset',
-  ]).isRequired as any,
-};
 
 export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   variant: ButtonVariants;
