--- packages/paste-core/components/button/tsconfig.build.json
@@ -1,42 +1,0 @@
-{
-  "extends": "../../../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "./src"
-  },
-  "include": [
-    "src/**/*"
-  ],
-  "references": [
-    {
-      "path": "../../primitives/box"
-    },
-    {
-      "path": "../../../paste-design-tokens"
-    },
-    {
-      "path": "../../../paste-icons"
-    },
-    {
-      "path": "../spinner"
-    },
-    {
-      "path": "../../layout/stack"
-    },
-    {
-      "path": "../../../paste-style-props"
-    },
-    {
-      "path": "../../../paste-libraries/styling"
-    },
-    {
-      "path": "../../../paste-theme"
-    },
-    {
-      "path": "../../../paste-types"
-    },
-    {
-      "path": "../../../paste-libraries/uid"
-    }
-  ]
-}
