+++ packages/paste-core/components/combobox/src/Combobox.tsx
@@ -7,15 +7,17 @@ import {ChevronDownIcon} from '@twilio-paste/icons/esm/ChevronDownIcon';
 import {Box} from '@twilio-paste/box';
 import {InputBox, InputChevronWrapper} from '@twilio-paste/input-box';
 import {Label} from '@twilio-paste/label';
-import {HelpText, HelpTextVariants} from '@twilio-paste/help-text';
-import {InputElement, InputVariants} from '@twilio-paste/input';
-import {TextColor} from '@twilio-paste/style-props';
+import {HelpText} from '@twilio-paste/help-text';
+import type {HelpTextVariants} from '@twilio-paste/help-text';
+import {InputElement} from '@twilio-paste/input';
+import {InputVariants} from '@twilio-paste/input';
+import type {TextColor} from '@twilio-paste/style-props';
 import {ComboboxInputWrapper} from './ComboboxInputWrapper';
 import {ComboboxListbox} from './ComboboxListbox';
 import {ComboboxListboxGroup} from './ComboboxListboxGroup';
 import {ComboboxListboxOption} from './ComboboxListboxOption';
-import {GroupItemsProps, ItemProps, ItemsProps, ListBoxProps, ComboboxProps} from './types';
 import {getIndexedItems, getGroupedItems} from './helpers';
+import type {GroupItemsProps, Item as ItemType, ItemProps, ItemsProps, ListBoxProps, ComboboxProps} from './types';
 
 // Fixes chevron overlapping really long text
 // Extra right padding is removed when autocomplete is true
