+++ packages/paste-core/components/combobox/src/index.tsx
@@ -1,10 +1,8 @@
-export {
-  useComboboxPrimitive as useCombobox,
-  UseComboboxPrimitiveState as UseComboboxState,
-} from '@twilio-paste/combobox-primitive';
+export {useComboboxPrimitive as useCombobox} from '@twilio-paste/combobox-primitive';
+export type {UseComboboxPrimitiveState as UseComboboxState} from '@twilio-paste/combobox-primitive';
 export * from './Combobox';
 export * from './ComboboxInputWrapper';
 export * from './ComboboxListbox';
 export * from './ComboboxListboxOption';
 export * from './ComboboxListboxGroup';
-export {ComboboxProps} from './types';
+export type {ComboboxProps} from './types';
