--- packages/paste-core/components/combobox/tsconfig.build.json
@@ -1,69 +1,0 @@
-{
-  "extends": "../../../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "./src"
-  },
-  "include": [
-    "src/**/*"
-  ],
-  "references": [
-    {
-      "path": "../../primitives/box"
-    },
-    {
-      "path": "../button"
-    },
-    {
-      "path": "../../primitives/combobox"
-    },
-    {
-      "path": "../../../paste-design-tokens"
-    },
-    {
-      "path": "../../layout/flex"
-    },
-    {
-      "path": "../help-text"
-    },
-    {
-      "path": "../../../paste-icons"
-    },
-    {
-      "path": "../input"
-    },
-    {
-      "path": "../input-box"
-    },
-    {
-      "path": "../label"
-    },
-    {
-      "path": "../screen-reader-only"
-    },
-    {
-      "path": "../spinner"
-    },
-    {
-      "path": "../../layout/stack"
-    },
-    {
-      "path": "../../../paste-style-props"
-    },
-    {
-      "path": "../../../paste-libraries/styling"
-    },
-    {
-      "path": "../../primitives/text"
-    },
-    {
-      "path": "../../../paste-theme"
-    },
-    {
-      "path": "../../../paste-types"
-    },
-    {
-      "path": "../../../paste-libraries/uid"
-    }
-  ]
-}
