+++ tsconfig.json
@@ -13,11 +13,11 @@
     // https://www.typescriptlang.org/docs/handbook/module-resolution.html#path-mapping
     "baseUrl": ".",
 
-    /* 
-    The lib is a special declaration file included by TS. It contains ambient declarations for 
-    common JS constructs that are present in runtimes and the DOM. Based on the target, TS 
-    automatically includes typings for the dom and the ES5 syntax. That’s why we need to 
-    specify ourselves we want typings for es2015 and the dom.  This way we have all the es6 
+    /*
+    The lib is a special declaration file included by TS. It contains ambient declarations for
+    common JS constructs that are present in runtimes and the DOM. Based on the target, TS
+    automatically includes typings for the dom and the ES5 syntax. That’s why we need to
+    specify ourselves we want typings for es2015 and the dom.  This way we have all the es6
     typings while targeting es5.
     */
     "lib": ["es2015", "dom", "scripthost", "es2015.promise"],
